+++ library/Oidc/CompatTabs.php
@@ -1,0 +1,14 @@
+<?php
+
+namespace Icinga\Module\Oidc;
+
+use ipl\Web\Widget\Tabs;
+
+class CompatTabs extends Tabs
+{
+    public function setTabs($tabs): Tabs
+    {
+        $this->tabs = $tabs;
+        return $this;
+    }
+}