+++ library/Oidc/Hook/GenericOidcImplementationHook.php
@@ -1,0 +1,8 @@
+<?php
+
+namespace Icinga\Module\Oidc\Hook;
+
+class GenericOidcImplementationHook
+{
+
+}