+++ library/Oidc/ProvidedHook/DefaultImplementation.php
@@ -1,0 +1,13 @@
+<?php
+
+/* originally from  Icinga Web 2 X.509 Module | (c) 2023 Icinga GmbH | GPLv2 */
+
+namespace Icinga\Module\Oidc\ProvidedHook;
+
+
+use Icinga\Module\Oidc\Hook\GenericOidcImplementationHook;
+
+class DefaultImplementation extends GenericOidcImplementationHook
+{
+
+}